pyproject.toml
[tool.poetry]
name = "convious-assignment"
version = "0.1.0"
description = ""
authors = ["dizvyagintsev <dizvyagintsev@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.9.0"
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
